add midleware

+++ routes/web.php
@@ -38,60 +38,58 @@ Route::prefix('admin')->middleware('auth')->group(function () {
     });
 
     //route dashboard
-    Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard');
+    Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard')->middleware('can:read-dashboard');
 
     //route master group
     Route::prefix('master')->group(function () {
         // Unit Produk
         Route::prefix('unit-produk')->group(function () {
-            Route::get('', [UnitProductController::class, 'index'])->name('unitproduk');
+            Route::get('', [UnitProductController::class, 'index'])->name('unitproduk')->middleware('can:read-unit-product');
             Route::get('getdata', [UnitProductController::class, 'GetData'])->name('unitproduk.getdata');
             Route::post('store', [UnitProductController::class, 'store'])->name('unitproduk.store');
             Route::get('/edit/{id}', [UnitProductController::class, 'show'])->name('unitproduk.edit');
-            Route::put('/update/{id}', [UnitProductController::class, 'update'])->name('unitproduk.update');
-            Route::delete('/delete/{id}', [UnitProductController::class, 'destroy'])->name('unitproduk.delete');
+            Route::put('/update/{id}', [UnitProductController::class, 'update'])->name('unitproduk.update')->middleware('can:update-unit-product');
+            Route::delete('/delete/{id}', [UnitProductController::class, 'destroy'])->name('unitproduk.delete')->middleware('can:delete-unit-product');
         });
 
         // Produk
         Route::prefix('produk')->group(function () {
-            Route::get('', [ProductController::class, 'index'])->name('produk');
+            Route::get('', [ProductController::class, 'index'])->name('produk')->middleware('can:read-product');
             Route::get('getdata', [ProductController::class, 'getData'])->name('produk.getdata');
             Route::post('store', [ProductController::class, 'store'])->name('produk.store');
-            Route::get('/edit/{id}', [ProductController::class, 'show'])->name('produk.edit');
+            Route::get('/edit/{id}', [ProductController::class, 'show'])->name('produk.edit')->middleware('can:update-product');
             Route::put('/update/{id}', [ProductController::class, 'update'])->name('produk.update');
-            Route::delete('/delete/{id}', [ProductController::class, 'destroy'])->name('produk.delete');
+            Route::delete('/delete/{id}', [ProductController::class, 'destroy'])->name('produk.delete')->middleware('can:delete-product');
         });
 
         // branch
         Route::prefix('cabang')->group(function () {
-            Route::get('', [BranchController::class, 'index'])->name('branch');
+            Route::get('', [BranchController::class, 'index'])->name('branch')->middleware('can:read-branch');;
             Route::get('getdata', [BranchController::class, 'getData'])->name('branch.getdata');
             Route::post('store', [BranchController::class, 'store'])->name('branch.store');
-            Route::get('/edit/{id}', [BranchController::class, 'show'])->name('branch.edit');
+            Route::get('/edit/{id}', [BranchController::class, 'show'])->name('branch.edit')->middleware('can:update-branch');
             Route::put('/update/{id}', [BranchController::class, 'update'])->name('branch.update');
-            Route::delete('/delete/{id}', [BranchController::class, 'destroy'])->name('branch.delete');
+            Route::delete('/delete/{id}', [BranchController::class, 'destroy'])->name('branch.delete')->middleware('can:delete-branch');
         });
 
         // zoneodp
         Route::prefix('jalur')->group(function () {
-            Route::get('', [ZoneOdpController::class, 'index'])->name('zone');
+            Route::get('', [ZoneOdpController::class, 'index'])->name('zone')->middleware('can:read-zone');
             Route::get('getdata', [ZoneOdpController::class, 'getData'])->name('zone.getdata');
             Route::get('syncdata', [ZoneOdpController::class, 'SyncData'])->name('zone.syncdata');
             Route::post('store', [ZoneOdpController::class, 'store'])->name('zone.store');
-            // Route::get('/edit/{id}', [ZoneOdpController::class, 'show'])->name('zone.edit');
-            // Route::put('/update/{id}', [ZoneOdpController::class, 'update'])->name('zone.update');
-            Route::delete('/delete/{id}', [ZoneOdpController::class, 'destroy'])->name('zone.delete');
+            Route::delete('/delete/{id}', [ZoneOdpController::class, 'destroy'])->name('zone.delete')->middleware('can:delete-zone');
         });
 
         //odp
         Route::prefix('odp')->group(function () {
-            Route::get('', [OdpController::class, 'index'])->name('odp');
+            Route::get('', [OdpController::class, 'index'])->name('odp')->middleware('can:read-zone-odp');
             Route::get('getdata', [OdpController::class, 'getData'])->name('odp.getdata');
             Route::get('syncdata', [OdpController::class, 'SyncData'])->name('odp.syncdata');
             Route::post('store', [OdpController::class, 'store'])->name('odp.store');
             // Route::get('/edit/{id}', [OdpController::class, 'show'])->name('odp.edit');
             // Route::put('/update/{id}', [OdpController::class, 'update'])->name('odp.update');
-            Route::delete('/delete/{id}', [OdpController::class, 'destroy'])->name('odp.delete');
+            Route::delete('/delete/{id}', [OdpController::class, 'destroy'])->name('odp.delete')->middleware('can:delete-zone-odp');
         });
 
         //product role
@@ -150,25 +148,25 @@ Route::prefix('admin')->middleware('auth')->group(function () {
         });
     });
 
-    //income product
-    Route::prefix('incomeproduct')->group(function () {
-        Route::get('', [IncomeProductController::class, 'index'])->name('incomeproduct');
-        Route::get('getdata', [IncomeProductController::class, 'getData'])->name('incomeproduct.getdata');
-        Route::post('store', [IncomeProductController::class, 'store'])->name('incomeproduct.store');
-        Route::get('/edit/{id}', [IncomeProductController::class, 'show'])->name('incomeproduct.edit');
-        Route::put('/update/{id}', [IncomeProductController::class, 'update'])->name('incomeproduct.update');
-        Route::delete('/delete/{id}', [IncomeProductController::class, 'destroy'])->name('incomeproduct.delete');
-    });
+    // //income product
+    // Route::prefix('incomeproduct')->group(function () {
+    //     Route::get('', [IncomeProductController::class, 'index'])->name('incomeproduct');
+    //     Route::get('getdata', [IncomeProductController::class, 'getData'])->name('incomeproduct.getdata');
+    //     Route::post('store', [IncomeProductController::class, 'store'])->name('incomeproduct.store');
+    //     Route::get('/edit/{id}', [IncomeProductController::class, 'show'])->name('incomeproduct.edit');
+    //     Route::put('/update/{id}', [IncomeProductController::class, 'update'])->name('incomeproduct.update');
+    //     Route::delete('/delete/{id}', [IncomeProductController::class, 'destroy'])->name('incomeproduct.delete');
+    // });
 
-    //outcome product
-    Route::prefix('outcomeproduct')->group(function () {
-        Route::get('', [OutcomeProductController::class, 'index'])->name('outcomeproduct');
-        Route::get('getdata', [OutcomeProductController::class, 'getData'])->name('outcomeproduct.getdata');
-        Route::post('store', [OutcomeProductController::class, 'store'])->name('outcomeproduct.store');
-        Route::get('/edit/{id}', [OutcomeProductController::class, 'show'])->name('outcomeproduct.edit');
-        Route::put('/update/{id}', [OutcomeProductController::class, 'update'])->name('outcomeproduct.update');
-        Route::delete('/delete/{id}', [OutcomeProductController::class, 'destroy'])->name('outcomeproduct.delete');
-    });
+    // //outcome product
+    // Route::prefix('outcomeproduct')->group(function () {
+    //     Route::get('', [OutcomeProductController::class, 'index'])->name('outcomeproduct');
+    //     Route::get('getdata', [OutcomeProductController::class, 'getData'])->name('outcomeproduct.getdata');
+    //     Route::post('store', [OutcomeProductController::class, 'store'])->name('outcomeproduct.store');
+    //     Route::get('/edit/{id}', [OutcomeProductController::class, 'show'])->name('outcomeproduct.edit');
+    //     Route::put('/update/{id}', [OutcomeProductController::class, 'update'])->name('outcomeproduct.update');
+    //     Route::delete('/delete/{id}', [OutcomeProductController::class, 'destroy'])->name('outcomeproduct.delete');
+    // });
 
     //settings
     Route::prefix('settings')->group(function () {
@@ -178,22 +176,22 @@ Route::prefix('admin')->middleware('auth')->group(function () {
         });
 
         Route::prefix('user')->group(function () {
-            Route::get('', [UserController::class, 'index'])->name('user');
+            Route::get('', [UserController::class, 'index'])->name('user')->middleware('can:read-user');
             Route::get('getdata', [UserController::class, 'getData'])->name('user.getdata');
             Route::post('store', [UserController::class, 'store'])->name('user.store');
-            Route::get('/edit/{id}', [UserController::class, 'show'])->name('user.edit');
+            Route::get('/edit/{id}', [UserController::class, 'show'])->name('user.edit')->middleware('can:update-user');
             Route::put('/update/{id}', [UserController::class, 'update'])->name('user.update');
-            Route::delete('/delete/{id}', [UserController::class, 'destroy'])->name('user.delete');
+            Route::delete('/delete/{id}', [UserController::class, 'destroy'])->name('user.delete')->middleware('can:delete-user');
         });
 
         Route::prefix('role')->group(function () {
-            Route::get('', [RoleController::class, 'index'])->name('role');
+            Route::get('', [RoleController::class, 'index'])->name('role')->middleware('can:read-role');
             Route::get('getdata', [RoleController::class, 'getData'])->name('role.getdata');
-            Route::get('add', [RoleController::class, 'create'])->name('role.add');
+            Route::get('add', [RoleController::class, 'create'])->name('role.add')->middleware('can:create-role');
             Route::post('store', [RoleController::class, 'store'])->name('role.store');
-            Route::get('/edit/{id}', [RoleController::class, 'show'])->name('role.edit');
+            Route::get('/edit/{id}', [RoleController::class, 'show'])->name('role.edit')->middleware('can:update-role');
             Route::put('/update/{id}', [RoleController::class, 'update'])->name('role.update');
-            Route::delete('/delete/{id}', [RoleController::class, 'destroy'])->name('role.delete');
+            Route::delete('/delete/{id}', [RoleController::class, 'destroy'])->name('role.delete')->middleware('can:delete-role');
         });
 
         Route::prefix('general')->group(function () {
